Overhauled the backend handling of kinematic variables. The change comes in a few pieces:
* Added ```set_kinematics``` function which allows the user to directly set the minimal kinematic variables needed to calculate the cross section.
* Changed ```read_data_file``` function to read data in line-by-line and create a list of pystruct dataclass instances available to the user.
* Removed the array index from cross section calculation related terms.

In addition, doc strings were added for most functions in the Femtogen and pymetric classes. Error checking added in some functions in preparation for unit testing in the near future. Lastly, an option to accept either a list of kinematic variables or as a pystruct dataclass added to ```set_kinematics```. Use the ```array``` keyword to access the pystruct option.